+++ src/components/select/select.styles.ts
@@ -68,7 +68,7 @@ export default css`
   }
 
   /* Visually hide the display input when multiple is enabled */
-  .select--multiple .select__display-input {
+  .select--multiple:not(.select--placeholder-visible) .select__display-input {
     position: absolute;
     z-index: -1;
     top: 0;
@@ -167,7 +167,7 @@ export default css`
     margin-inline-end: var(--sl-input-spacing-small);
   }
 
-  .select--small.select--multiple .select__combobox {
+  .select--small.select--multiple:not(.select--placeholder-visible) .select__combobox {
     padding-block: 2px;
     padding-inline-start: 0;
   }
@@ -192,7 +192,7 @@ export default css`
     margin-inline-end: var(--sl-input-spacing-medium);
   }
 
-  .select--medium.select--multiple .select__combobox {
+  .select--medium.select--multiple:not(.select--placeholder-visible) .select__combobox {
     padding-inline-start: 0;
     padding-block: 3px;
   }
@@ -217,7 +217,7 @@ export default css`
     margin-inline-end: var(--sl-input-spacing-large);
   }
 
-  .select--large.select--multiple .select__combobox {
+  .select--large.select--multiple:not(.select--placeholder-visible) .select__combobox {
     padding-inline-start: 0;
     padding-block: 4px;
   }